[Object] Widen data size calculation to 64-bits to prevent overflow
This patch widens selected operands of the product calculation to
`size_t`/`uint64_t` to prevent multiplication overflow when a 64-bit
data size or offset are being calculated. This fixes a number of
issues reported by a CodeQL scan.

+++ llvm/include/llvm/Object/DXContainer.h
@@ -210,7 +210,8 @@ struct DescriptorTableView : RootParameterView {
     Current += sizeof(uint32_t);
 
     Table.Ranges.Data = ParamData.substr(2 * sizeof(uint32_t),
-                                         Table.NumRanges * Table.Ranges.Stride);
+                                         static_cast<size_t>(Table.NumRanges) *
+                                             Table.Ranges.Stride);
     return Table;
   }
 };
